requirements.txt
# LangChain core
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0

# Vector store
chromadb>=0.4.22

# Embeddings
tiktoken>=0.5.2

# Document loaders
pypdf>=3.17.0
unstructured>=0.11.0

# Web UI
streamlit>=1.29.0

# REST API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Web scraping (for loading web content)
beautifulsoup4>=4.12.0
requests>=2.31.0
